tracking: move multi-camera people tracking example into new category
Create a first-class tracking example category and move the multi-camera
people tracking example under it with an OptiView-specific slug.

- add examples/tracking and register it in the examples build
- move the example to tracking/multi-camera-people-detection-and-tracking-optiview
- rename the example binary and C++ test targets to use the -optiview suffix
- update C++ internal include paths and default config path after the move
- refresh README metadata, title, commands, and source file references
- allow tracking as a valid README category
- remove the leftover common/.gitkeep placeholder
