--- examples/object-detection/multi-camera-people-detection-and-tracking/common/.gitkeep
@@ -1,1 +1,0 @@
-
