+++ examples/tracking/CMakeLists.txt
@@ -1,7 +1,5 @@
 foreach(example IN ITEMS
-    single-rtsp-object-detection-optiview
-    multistream-rtsp-detection-pipeline
-    simple-object-detection-overlay-pipeline
+    multi-camera-people-detection-and-tracking-optiview
 )
   if(EXISTS "${CMAKE_CURRENT_SOURCE_DIR}/${example}/cpp/CMakeLists.txt")
     add_subdirectory(${example}/cpp ${example})
