+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/common/config.yaml
