+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/config_api.cpp
