+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/tests/CMakeLists.txt
