+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/tests/e2e_test.cpp
