+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/cpp/tracker_api.cpp
