+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/python/requirements.txt
