+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/python/tests/test_workers.py
