+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/python/utils/__init__.py
