+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/python/utils/image_utils.py
