+++ examples/tracking/multi-camera-people-detection-and-tracking-optiview/python/utils/sample_utils.py
